Cleanup after refactor
Manual removal of netowrk is not necessary, testcontainers automatically
cleans up afterwards. And with multiple tests classes the network will be
deleted after the first class, while next ones attempt to use it.

Also as the vault container is declared once, but can be started multiple
times, need to store the startup commands beyond first startup.