Extract the repository location outside the script
Extract the repository location to a location outside the hardcoded
JobDSL script. Will save as a basis for later parametrisation.

Had to change the hardcoded script from static to an instance variable,
as while Groovy runs the closure when the GString is converted to
string, it binds it earlier, and there's no property at static level.