Allow for logging in via wrapped secret ID
Vault allows for wrapping response in single-use tokens. The token can be
passed around freely, but it's only possible to read it once. This provides
additional layer of protection, as the infrastructure cannot covertly
read the response.

Implemented a mechanism that instead of passing AppRole Role ID and Secret
ID directly allows passing a response token that wraps the latter.